+++ src/main/java/ackermann/Ackermann.java
@@ -1,9 +1,14 @@
+package ackermann;
+
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Scanner;
 
-import exceptions.CheckedException;
-import mypackage.*;
+import ackermann.exceptions.CheckedException;
+import ackermann.ui.Parser;
+import ackermann.ui.Storage;
+import ackermann.ui.TaskList;
+import ackermann.ui.Ui;
 
 public class Ackermann {
     private static final String LINE = "-----------------------------------------------------";
