+++ src/main/java/ackermann/exceptions/task/InvalidDeadline/InvalidDeadlineByException.java
@@ -1,4 +1,4 @@
-package exceptions.task.InvalidDeadline;
+package ackermann.exceptions.task.InvalidDeadline;
 
 public class InvalidDeadlineByException extends InvalidDeadlineException{
