+++ src/main/java/ackermann/task/Events.java
@@ -1,4 +1,4 @@
-package mypackage.task;
+package ackermann.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
