+++ src/main/java/ackermann/task/ToDos.java
@@ -1,4 +1,4 @@
-package mypackage.task;
+package ackermann.task;
 
 public class ToDos extends Task {
     public ToDos(String name) {
