+++ src/main/java/ackermann/ui/Ui.java
@@ -1,15 +1,15 @@
-package mypackage;
+package ackermann.ui;
 
-import exceptions.InvalidTargetException;
-import exceptions.task.InvalidDeadline.InvalidDeadlineByException;
-import exceptions.task.InvalidDeadline.InvalidDeadlineException;
-import exceptions.task.InvalidEvent.InvalidEventException;
-import exceptions.task.InvalidEvent.InvalidEventFromException;
-import exceptions.task.InvalidEvent.InvalidEventToException;
-import mypackage.task.Deadlines;
-import mypackage.task.Events;
-import mypackage.task.Task;
-import mypackage.task.ToDos;
+import ackermann.exceptions.InvalidTargetException;
+import ackermann.exceptions.task.InvalidDeadline.InvalidDeadlineByException;
+import ackermann.exceptions.task.InvalidDeadline.InvalidDeadlineException;
+import ackermann.exceptions.task.InvalidEvent.InvalidEventException;
+import ackermann.exceptions.task.InvalidEvent.InvalidEventFromException;
+import ackermann.exceptions.task.InvalidEvent.InvalidEventToException;
+import ackermann.task.Deadlines;
+import ackermann.task.Events;
+import ackermann.task.Task;
+import ackermann.task.ToDos;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
